Create-project (#459)
* Fixed compilation error

* Added the creation of project handler in Avalonia

* Fixed issues with the create project on Avalonia

* Fixed the broken tests after method signature change

* Fixed the wallet issue in the create project page

* Fixed percentage issue

* Uncommented the indexer

+++ src/Angor/Avalonia/AngorApp/Sections/Founder/CreateProject/Stages/ICreateProjectStage.cs
@@ -5,7 +5,7 @@ namespace AngorApp.Sections.Founder.CreateProject.Stages;
 
 public interface ICreateProjectStage : IReactiveObject, IValidatableViewModel
 {
-    public double? Percent { get; set; }
+    public decimal? Percent { get; set; }
     public DateTime? ReleaseDate { get; set; }
     public IEnhancedCommand Remove { get; }
 }
